test(mpsc): organize coverage around channel contracts
